split ntrip.go into source_table.go, too
